requirements.txt
requests
python-dotenv
dnspython
python-whois
colorama
Flask